massive refactor of the repo structure and added color pickers
