+++ app_src/shiny_server.py
@@ -1,0 +1,437 @@
+"""
+This module creates the Server logic
+"""
+
+# libs
+from shiny import render, ui, reactive
+import matplotlib.pyplot as plt
+
+# load in app resources
+from app_src.shiny_plots import set_aln, create_msa_plot, create_analysis_custom_heatmap, create_freq_heatmap
+from shinywidgets import render_widget
+
+# msaexplorer
+from msaexplorer import explore, config, export
+
+
+def server(input, output, session):
+    """
+    creates the server logic
+    """
+
+    reactive.alignment = reactive.Value(None)
+    reactive.annotation = reactive.Value(None)
+
+    # create inputs for plotting and pdf
+    def prepare_inputs():
+        """
+        Collect inputs from the UI for the plot tab, only adding the ones needed for enabled features.
+        This ensures unnecessary reactivity.
+        """
+        inputs = {}
+        aln = reactive.alignment.get()
+
+        if aln is None:
+            return None
+
+        # inhibit the window dimensions and height to trigger a re-rendering
+        # as this is only need for the calc whether to show the
+        # sequence but itself does not change the plots appearance
+        # --> other reactive values will change anyway
+        with reactive.isolate():
+            dims = input.window_dimensions()
+            window_width = dims['width']
+            window_height = dims['height']
+            increase_height = input.increase_height()
+
+        # Always needed inputs
+        inputs['zoom_range'] = input.zoom_range()
+        inputs['reference'] = input.reference()
+        inputs['alignment_type'] = input.alignment_type()
+        inputs['annotation'] = input.annotation()
+        inputs['stat_type'] = input.stat_type()
+
+        # STATISTICS (first plot)
+        if inputs['stat_type'] != 'Off':
+            inputs['plot_1_size'] = input.plot_1_size()
+            inputs['rolling_average'] = input.rolling_avg()
+            inputs['stat_color'] = input.stat_color()
+
+        # ALIGNMENT (second plot)
+        if inputs['alignment_type'] != 'Off':
+            inputs['reference_color'] = input.reference_color()
+            inputs['plot_2_size'] = input.plot_2_size()
+            inputs['fancy_gaps'] = input.fancy_gaps()
+            inputs['show_mask'] = input.show_mask()
+            inputs['show_gaps'] = input.show_gaps()
+            inputs['show_legend'] = input.show_legend()
+            inputs['show_ambiguities'] = input.show_ambiguities()
+            if inputs['alignment_type'] not in ['identity', 'colored identity']:
+                inputs['matrix'] = input.matrix()
+                inputs['matrix_color_mapping'] = input.matrix_color_mapping()
+
+            # determine if it makes sense to show the sequence or sequence names
+            # therefore figure out if there are enough chars/size that sequence fits in there
+            complete_size = input.plot_2_size()
+            if inputs['stat_type'] != 'Off':
+                complete_size += input.plot_1_size()
+            if inputs['annotation'] != 'Off':
+                complete_size += input.plot_3_size()
+            relative_msa_height = input.plot_2_size() * increase_height / complete_size * window_height / len(aln.alignment)
+            relative_msa_width = window_width / (inputs['zoom_range'][1]-inputs['zoom_range'][0])
+            # and then decide how to set the show sequence input
+            if relative_msa_width >= 11 and relative_msa_height >= 18:
+                inputs['show_sequence'] = input.show_sequence()
+            else:
+                inputs['show_sequence'] = False
+            # and the seq_names input - check if there is enough y space for the text
+            if relative_msa_height >= 15:
+                inputs['seq_names'] = input.seq_names()
+            else:
+                inputs['seq_names'] = False
+
+        # ANNOTATION (third plot)
+        if inputs['annotation'] != 'Off':
+            inputs['plot_3_size'] = input.plot_3_size()
+            if inputs['annotation'] == 'Annotation':
+                inputs['annotation_file'] = input.annotation_file()
+                inputs['feature_display'] = input.feature_display()
+                inputs['feature_color'] = input.feature_color()
+                inputs['strand_marker_size'] = input.strand_marker_size()
+            elif inputs['annotation'] == 'Conserved ORFs':
+                inputs['color_mapping'] = input.color_mapping()
+                inputs['non_overlapping'] = input.non_overlapping()
+                inputs['min_orf_length'] = input.min_orf_length()
+                inputs['strand_marker_size'] = input.strand_marker_size()
+                inputs['show_legend_third_plot'] = input.show_legend_third_plot()
+            else:
+                inputs['stem_size'] = input.stem_size()
+                inputs['head_size'] = input.head_size()
+                inputs['show_legend_third_plot'] = input.show_legend_third_plot()
+
+        return inputs
+
+    # separate function if not all inputs are needed
+    def prepare_minimal_inputs(zoom: bool = True, window_size: bool = False, ref: bool = False, plot: bool = False):
+        """
+        minimal inputs with options depending on which are needed
+        """
+        inputs = {}
+        aln = reactive.alignment.get()
+
+        if aln is None:
+            return None
+
+        if zoom:
+            inputs['zoom_range'] = input.zoom_range()
+        if ref:
+            inputs['reference'] = input.reference()
+        if plot:
+            inputs['analysis_plot_type'] = input.analysis_plot_type()
+            if inputs['analysis_plot_type'] != 'Off':
+                inputs['additional_analysis_options'] = input.additional_analysis_options()
+        if window_size:
+            inputs['dimensions'] = input.window_dimensions()
+
+        return inputs
+
+    def read_in_annotation(annotation_file):
+        """
+        Read in an annotation and update the ui accordingly
+        """
+        ann = explore.Annotation(reactive.alignment.get(), annotation_file[0]['datapath'])
+        reactive.annotation.set(ann)
+
+        # update features to display
+        ui.update_selectize(
+            id='feature_display',
+            choices=list(ann.features.keys()),
+            selected=list(ann.features.keys())[0]
+        )
+
+        # update possible user inputs
+        if reactive.alignment.get().aln_type == 'AA':
+            ui.update_selectize('annotation', choices=['Off', 'SNPs', 'Annotation'], selected='Annotation')
+        else:
+            ui.update_selectize('annotation', choices=['Off', 'SNPs', 'Conserved ORFs', 'Annotation'],
+                                selected='Annotation')
+
+    # Updates the plot container
+    @reactive.Effect
+    async def update_height():
+        """
+        update the plot container height -> Sends a message that is picked up by the js and updates the CSS height
+        property for the plot container
+        """
+        new_plot_height = f'{input.increase_height() * 100}vh'
+
+        await session.send_custom_message("update-plot-container-height", {'height': new_plot_height})
+
+    # Inputs
+    @reactive.Effect
+    @reactive.event(input.alignment_file)
+    def load_alignment():
+        """
+        Load an alignment and update different now accessible ui elements
+        """
+        try:
+            alignment_file = input.alignment_file()
+            annotation_file = input.annotation_file()
+            if alignment_file:
+                aln = explore.MSA(alignment_file[0]['datapath'], reference_id=None, zoom_range=None)
+
+                # set standard ref
+                aln.reference_id = list(aln.alignment.keys())[0]
+                reactive.alignment.set(aln)
+
+                # Update zoom slider based on alignment length and user input
+                alignment_length = len(next(iter(aln.alignment.values())))-1
+                ui.update_slider('zoom_range', max=alignment_length-1, value=(0, alignment_length -1))
+
+                # Update reference
+                ui.update_selectize(
+                    id='reference', choices=['first', 'consensus'] + list(aln.alignment.keys()), selected='first'
+                )
+
+                # Update substitution matrix
+                ui.update_selectize(
+                    id='matrix',
+                    choices=list(config.SUBS_MATRICES[aln.aln_type].keys()),
+                    selected='BLOSUM65' if aln.aln_type == 'AA' else 'TRANS',
+                )
+
+                # update plot size sliders
+                # Adjust the size depending on the number of alignment sequences
+                aln_len, seq_threshold = len(aln.alignment.keys()), 5
+                for ratio in config.STANDARD_HEIGHT_RATIOS.keys():
+                    if aln_len >= ratio:
+                        seq_threshold = ratio
+
+                # update standard settings
+                ui.update_numeric('plot_1_size', value=config.STANDARD_HEIGHT_RATIOS[seq_threshold][0])
+                ui.update_numeric('plot_2_size', value=config.STANDARD_HEIGHT_RATIOS[seq_threshold][1])
+                ui.update_numeric('plot_3_size', value=config.STANDARD_HEIGHT_RATIOS[seq_threshold][2])
+                # Some of the function highly depend on the alignment type
+                if aln.aln_type == 'AA':
+                    ui.update_selectize('stat_type', choices=['Off', 'entropy', 'coverage', 'identity', 'similarity'], selected='Off')
+                    ui.update_selectize('annotation', choices=['Off', 'SNPs'])
+                else:
+                    # needed because if an as aln and then a nt aln are loaded it will not change
+                    ui.update_selectize('stat_type', choices=['Off', 'gc', 'entropy', 'coverage', 'identity', 'similarity', 'ts tv score'], selected='Off')
+                    ui.update_selectize('annotation', choices=['Off', 'SNPs', 'Conserved ORFs'])
+                # case if annotation file is uploaded prior to the alignment file
+                if annotation_file:
+                    read_in_annotation(annotation_file)
+        # show the user if something with parsing went wrong
+        except Exception as e:
+            print(f"Error: {e}")  # print to console
+            ui.notification_show(ui.tags.div(
+                    f'Error: {e}',
+                    style="color: red; font-weight: bold;"
+                ), duration=10)  # print to user
+
+    @reactive.Effect
+    @reactive.event(input.annotation_file)
+    def load_annotation():
+        """
+        load the annotation - catches errors if its the wrong format and displays it (otherwise app_src would crash)
+        """
+        # read annotation file
+        try:
+            annotation_file, alignment_file = input.annotation_file(), input.alignment_file()
+            if annotation_file and alignment_file:
+                read_in_annotation(annotation_file)
+        # show the user if something with parsing went wrong
+        except Exception as e:
+            print(f"Error: {e}")  # print to console
+            ui.notification_show(ui.tags.div(
+                f'Error: {e}',
+                style="color: red; font-weight: bold;"
+            ), duration=10)  # print to user
+
+
+    @reactive.Effect
+    @reactive.event(input.download_type)
+    def update_download_options():
+        """
+        Update the UI - which download options are displayed
+        """
+        if input.download_type() == 'SNPs':
+            ui.update_selectize('download_format', choices=['vcf', 'tabular'])
+
+    @render.download()
+    def download_stats():
+        """
+        Download various files in standard format
+        """
+        try:
+            # Initialize
+            download_format = input.download_format()
+            aln = reactive.alignment.get()
+            if aln is None:
+                raise FileNotFoundError("No alignment data available. Please upload an alignment.")
+
+            if input.download_type() == 'SNPs':
+                download_data = export.snps(aln.get_snps(), format_type=download_format)
+                prefix = 'SNPs_'
+
+            # Create a temporary file for the download
+            with tempfile.NamedTemporaryFile(prefix=prefix, suffix=f'.{download_format}', delete=False) as tmpfile:
+                tmpfile.write(download_data.encode('utf-8'))
+                tmpfile.flush()  # Ensure data is written to disk
+
+                return tmpfile.name
+        # also send a notification to the user
+        except FileNotFoundError:
+            ui.notification_show(ui.tags.div(
+                'No alignment was uploaded.',
+                style="color: red; font-weight: bold;"
+            ), duration=10)
+
+    @output
+    @render.plot
+    def msa_plot():
+        """
+        plot the alignment
+        """
+        aln = reactive.alignment.get()
+        ann = reactive.annotation.get()
+
+        return create_msa_plot(aln, ann, prepare_inputs())
+
+    @output
+    @render.download
+    def download_pdf():
+        """
+        allows the export of the msa plot as pdf - importantly it will be same size as the
+        window ensuring that the plot is exactly the same as the rendered displayed plot
+        """
+        # get annotation and alignment
+        aln = reactive.alignment.get()
+        ann = reactive.annotation.get()
+
+        # access the window dimensions
+        dimensions = input.window_dimensions_plot()
+        figure_width_inches = dimensions['width'] / 96
+        figure_height_inches = dimensions['height'] / 96 * input.increase_height()
+
+        # plot with a temp name
+        with tempfile.NamedTemporaryFile(suffix=".pdf", delete=False) as tmpfile:
+            fig = create_msa_plot(aln, ann, prepare_inputs(), fig_size=(figure_width_inches, figure_height_inches))
+            # tight layout needed here to plot everything correctly
+            fig.tight_layout()
+            fig.savefig(tmpfile.name, format="pdf")
+            plt.close(fig)
+            return tmpfile.name
+
+    # showcases:
+    @render.ui
+    def aln_type():
+        aln = reactive.alignment.get()
+        if aln is None:
+            return None
+        return aln.aln_type
+
+    @render.ui
+    def zoom_range_analysis():
+        aln = reactive.alignment.get()
+        if aln is None:
+            return None
+
+        aln = set_aln(aln, prepare_minimal_inputs())
+
+        return f'{aln.zoom[0]} - {aln.zoom[1]}'
+
+    @render.ui
+    def number_of_seq():
+        aln = reactive.alignment.get()
+        if aln is None:
+            return None
+
+        return len(aln.alignment)
+
+    @render.ui
+    def aln_len():
+        aln = reactive.alignment.get()
+        if aln is None:
+            return None
+
+        aln = set_aln(aln, prepare_minimal_inputs())
+
+        return aln.length
+
+    @render.ui
+    def per_gaps():
+        aln = reactive.alignment.get()
+        if aln is None:
+            return None
+
+        aln = set_aln(aln, prepare_minimal_inputs())
+
+        return round(aln.calc_character_frequencies()['total']['-']['% of alignment'], 2)
+
+    @render.ui
+    def snps():
+        aln = reactive.alignment.get()
+        if aln is None:
+            return None
+
+        aln = set_aln(aln, prepare_minimal_inputs(ref=True))
+
+        return len(aln.get_snps()['POS'])
+
+    @reactive.Effect
+    @reactive.event(input.analysis_plot_type)
+    def update_additional_options():
+        """
+        Update UI for the analysis tab
+        """
+        # ensure that it is switched back
+        if input.analysis_plot_type() == 'Off':
+            ui.update_selectize('additional_analysis_options', choices=['None'], selected='None')
+        if input.analysis_plot_type() == 'Pairwise identity':
+            ui.update_selectize(
+                'additional_analysis_options',
+                choices={
+                    'ghd': 'global hamming distance',
+                    'lhd': 'local hamming distance',
+                    'ged': 'gap excluded distance',
+                    'gcd': 'gap compressed distance'
+                },
+                selected='ghd'
+            )
+
+    @render.text
+    def analysis_info():
+        """
+        show custom text for additional options
+        """
+        selected_option = input.additional_analysis_options()
+        if selected_option == "ghd":
+            return 'INFO ghd (global hamming distance):\n\nAt each alignment position, check if\ncharacters match:\n\ndistance = matches / alignment_length * 100'
+        elif selected_option == "lhd":
+            return 'INFO lhd (local hamming distance):\n\nRestrict the alignment to the region\nin both sequences that do not start\nand end with gaps:\n\ndistance = matches / min(end-ungapped seq1, end-ungapped seq2) * 100'
+        elif selected_option == 'ged':
+            return 'INFO ged (gap excluded distance):\n\nAll gaps are excluded from the \nalignment\n\ndistance = matches / (matches + mismatches) * 100'
+        elif selected_option == 'gcd':
+            return 'INFO gcd (gap compressed distance):\n\nAll consecutive gaps arecompressed to\none mismatch.\n\ndistance = matches / gap_compressed_alignment_length * 100'
+        else:
+            return None
+
+
+    @render_widget
+    def analysis_custom_heatmap():
+        """
+        Create the heatmap
+        """
+        aln = reactive.alignment.get()
+
+        return create_analysis_custom_heatmap(aln, prepare_minimal_inputs( plot=True, window_size=True))
+
+
+    @render_widget
+    def analysis_char_freq_heatmap():
+        aln = reactive.alignment.get()
+
+        return create_freq_heatmap(aln,  prepare_minimal_inputs(window_size=True))
+
